Seperated CMAKE files and fixed MSVC compiler warnings
Seperatd the Cmake files into the configuration and the code. This makes
it easier to seperate compiler configuration and cmake configuration
from the source files/needed libraries.
Also added UWS_SHARED  pre processor define when the project is built
for windows and BUILD_SHARED is true. When building a lib and
__declspec(dllexport) windows VS issues a bunch of warnings. This patch
removes the warnings by only using  __declspec(dllexport) when building
a shared library.